// sma-contact-flow/sma-contact-flow-parser/src/const/error-types.ts
// Error types that SMA will return if any of the action gets failed
export enum ErrorTypes {
    INPUT_TIME_LIMIT_EXCEEDS = "InputTimeLimitExceeded",
    INVALID_DIGITS_RECEIVED = "InvalidDigitsReceived",
    NO_MATCHING_CONDITION = "NoMatchingCondition",
    CONNECTION_TIME_LIMIT_EXCEEDED = "ConnectionTimeLimitExceeded",
    CALL_FAILED = "CallFailed",
    INVALID_PHONE_NUMBER = "InvalidPhoneNumber",
    NO_MATCHING_ERROR = "NoMatchingError"
}